% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{extractData}
\alias{extractData}
\title{Get data from a csv file}
\usage{
extractData(filename, errorbars = FALSE)
}
\arguments{
\item{filename}{Self-explanatory.}
}
\description{
This function retrieves and pre-processes the dataset from the given file into a suitable format for oscillock.
}
\examples{
cat_function()
}
