% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscDetection.R
\name{oscOrNot}
\alias{oscOrNot}
\title{Oscillatory or not?}
\usage{
oscOrNot(
  d,
  threshold = 14,
  detrend = TRUE,
  nonstat = TRUE,
  plotting = FALSE,
  Q = 1
)
}
\arguments{
\item{d}{The dataset to be tested, a list with features X (time points), Y (averaged replicates), S (error bars).}

\item{threshold}{The cutoff value for the bayes factor for classifying oscillatory data. Defaults to 14.}

\item{detrend}{Do you want to detrend the data via linear regression? Defaults to TRUE. If your data is already detrended OR you do not want to detrend the data at all, set this parameter to FALSE.}

\item{nonstat}{Do you want to use the nonstationary kernel? Defaults to TRUE.}

\item{plotting}{Do you want a plot of the optimised non-stationary kernel posterior on the input data? Defaults to FALSE.}
}
\description{
This function is the main interface function that is called to determine whether a dataset is oscillatory or not.
}
\examples{
test_function()
}
